+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## 2023.07.10 v5.2.0
 
-1. new `awaitTo` `fingerprint` function
+1. new `awaitTo` `fingerprint` `browserVersion` function
 2. Refactored `client` method
-3. fix bugs for `osVersion`
+3. fix bugs of `osVersion`
 4. deprecated `trim`
 5. docs work